+++ ui/great-building.js
@@ -41,18 +41,24 @@ function updateGreatBuildingBoostInfo (fpAnalysis) {
 }
 
 function updateGreatBuildingChanges (changes) {
-  $('#great-building-changes-player').text(changes.player)
+  $('#great-building-changes-player').text(changes.player.name)
+
+  let playerInfoHTML
+  if (changes.player.info === undefined || changes.player.info.lastUpdate === undefined) {
+    playerInfoHTML = '<span style="color:orange;">no info</span>'
+  } else {
+    playerInfoHTML = (changes.player.info.active ? 'active' : '<span style="color:red;">inactive</span>')
+    playerInfoHTML += ' <span style="font-size:8pt">updated ' + moment.unix(changes.player.info.lastUpdate).fromNow(false) + '</span>'
+  }
+  $('#great-building-changes-player-info').html(playerInfoHTML)
+
   changes = changes.changes
   let ggRows = ''
   if (changes.length === 0) {
     ggRows = '<tr><td colspan="3">No buildings</td></tr>'
   } else {
-    if (changes.length === 1 && changes[0].name === 'last change') {
-      ggRows = '<tr><td colspan="3">' + (changes[0].last_spent ? 'Last change ' + moment.unix(changes[0].last_spent).fromNow() : 'Not available') + '</td></tr>'
-    } else {
-      for (let i = 0; i < changes.length; i++) {
-        ggRows += addGreatBuildingChangesRow(changes[i])
-      }
+    for (let i = 0; i < changes.length; i++) {
+      ggRows += addGreatBuildingChangesRow(changes[i])
     }
   }
 
@@ -60,7 +66,7 @@ function updateGreatBuildingChanges (changes) {
 }
 
 function addGreatBuildingAnalysisRow (spot, analysis) {
-  let row = '<tr><td>' + spot + '</td><td>' + iconImage('sp') + ' ' + analysis.spotSafe + (analysis.spotSafe <= 0 ? ' (safe)' : '') + '</td><td' + (analysis.spotSafe >= 0 ? (analysis.profit < 0 ? ' style="color:red;"' : '') + '>' + iconImage('sp') + ' ' + analysis.profit : '>') + '</td>'
+  let row = '<tr><td>' + spot + '</td><td>' + iconImage('strategy_points') + ' ' + analysis.spotSafe + (analysis.spotSafe <= 0 ? ' (safe)' : '') + '</td><td' + (analysis.spotSafe >= 0 ? (analysis.profit < 0 ? ' style="color:red;"' : '') + '>' + iconImage('strategy_points') + ' ' + analysis.profit : '>') + '</td>'
   row += '<td>'
   if (analysis.reward.blueprints) {
     row += iconImage('blueprint') + ' ' + analysis.reward.blueprints
@@ -71,7 +77,7 @@ function addGreatBuildingAnalysisRow (spot, analysis) {
   }
   row += '</td><td>'
   if (analysis.reward.medals) {
-    row += iconImage('medal') + ' ' + analysis.reward.medals
+    row += iconImage('medals') + ' ' + analysis.reward.medals
   }
   row += '</td><td>'
   if (analysis.reward.medalsBonus) {
@@ -79,7 +85,7 @@ function addGreatBuildingAnalysisRow (spot, analysis) {
   }
   row += '</td><td>'
   if (analysis.reward.fp) {
-    row += iconImage('sp') + ' ' + analysis.reward.fp
+    row += iconImage('strategy_points') + ' ' + analysis.reward.fp
   }
   row += '</td><td>'
   if (analysis.reward.fpBonus) {
@@ -103,7 +109,7 @@ function addGreatBuildingBoostRow (spot, boostFactor, analysis, totalFP, freeFP,
 
   let row = '<tr' + (unsafe ? ' class="table-danger"' : (notInvestedEnough ? ' class="table-warning"' : '')) + '>'
   row += '<td>' + spot + '</td>'
-  row += '<td>' + iconImage('sp') + ' ' + required + (notInvestedEnough ? ` (${analysis.invested})` : '') + '</td><td>' + iconImage('sp') + ' ' + fillForSafe + (unsafe ? ` (${requiredToMakeSafe})` : '') + '</td>'
+  row += '<td>' + iconImage('strategy_points') + ' ' + required + (notInvestedEnough ? ` (${analysis.invested})` : '') + '</td><td>' + iconImage('strategy_points') + ' ' + fillForSafe + (unsafe ? ` (${requiredToMakeSafe})` : '') + '</td>'
   row += '<td>'
   if (analysis.reward.blueprints) {
     row += iconImage('blueprint') + ' ' + analysis.reward.blueprints
@@ -114,7 +120,7 @@ function addGreatBuildingBoostRow (spot, boostFactor, analysis, totalFP, freeFP,
   }
   row += '</td><td>'
   if (analysis.reward.medals) {
-    row += iconImage('medal') + ' ' + analysis.reward.medals
+    row += iconImage('medals') + ' ' + analysis.reward.medals
   }
   row += '</td><td>'
   if (analysis.reward.medals) {
@@ -125,5 +131,5 @@ function addGreatBuildingBoostRow (spot, boostFactor, analysis, totalFP, freeFP,
 }
 
 function addGreatBuildingChangesRow (change) {
-  return '<tr><td>' + change.name + '</td><td>' + (change.last_spent ? moment.unix(change.last_spent).fromNow(true) : '<em>closed</em>') + '</td><td>' + change.completePercentage + '%</td></tr>'
+  return '<tr><td>' + change.name + '</td><td>' + (change.lastChange !== undefined ? moment.unix(change.lastChange).fromNow(true) : '<em>???</em>') + '</td><td>' + change.completePercentage + '%</td></tr>'
 }